Tema 3/requirements.txt
Flask==1.0.2
gunicorn==19.9.0
google-cloud-storage==1.13.2
google-cloud-datastore==1.7.3
grpcio == 1.19.0
